refactor: rename zerotouch paths to autopilot
"Autopilot" is the established product name for this rewrite (already used
in README.md, docs/design/v3-rollout-plan.md, and several docs/content/legacy
pages) -- "zero-touch" was this session's first guess before that surfaced.
Renames:

- packages/@cdklabs/cdk-cicd-wrapper-cli/src/cmds/zerotouch/ -> cmds/autopilot/
- packages/@cdklabs/cdk-cicd-wrapper-cli/test/zerotouch/ -> test/autopilot/
- docs/content/workshops/zerotouch-pipeline/ -> docs/content/workshops/autopilot-pipeline/

Updated every import path in src/index.ts and the moved test files.

Verified: cdk-cicd-wrapper-cli compiles clean and its full test suite
passes (9/9 suites, 112/112 tests) against the new paths.
