+++ docs/content/workshops/autopilot-pipeline/.pages.yml
