+++ docs/content/workshops/autopilot-pipeline/00-prerequisites.md
