+++ docs/content/workshops/autopilot-pipeline/01-config-driven-pipeline.md
@@ -5,7 +5,7 @@
     - An understanding of **every** `defineCICD` field and when to reach for it.
     - A customized CI phase — your own named build steps and, optionally, a custom build image.
 
-The whole v3 opt-in is two things: a `cicd.config.ts`, and pointing `cdk.json` at `cdk-cicd exec`.
+The whole zero-touch opt-in is two things: a `cicd.config.ts`, and pointing `cdk.json` at `cdk-cicd exec`.
 
 ## 1. Your app stays plain CDK
 
@@ -60,9 +60,9 @@ This deploys **one** pipeline into your hub account from `cicd.config.ts` alone.
 Source → Build (checks + synth) → UpdatePipeline (re-deploys itself from config) → deploy dev → deploy prod
 ```
 
-![The flat v3 pipeline in the CodePipeline console: Source → Build → UpdatePipeline → deploy](images/flat-pipeline.png)
+![The flat zero-touch pipeline in the CodePipeline console: Source → Build → UpdatePipeline → deploy](images/flat-pipeline.png)
 
-*The flat v3 pipeline in the CodePipeline console — one linear pipeline, not Blueprint's 100+ CodeBuild projects.*
+*The flat zero-touch pipeline in the CodePipeline console — one linear pipeline, not Blueprint's 100+ CodeBuild projects.*
 
 The **UpdatePipeline** stage means you never run `deploy-ci` again by hand: change `cicd.config.ts`, push,
 and the pipeline re-synthesizes its own definition on the next run and applies the change before the
@@ -145,7 +145,7 @@ chapter 3.
     - In the **CodePipeline** console, your pipeline shows the flat
       `Source → Build → UpdatePipeline → deploy dev → deploy prod` shape.
     - The most recent execution reaches **Succeeded** on every stage.
-    - Compare the CodeBuild project count to a Blueprint pipeline: v3 provisions a small, constant set — not one
+    - Compare the CodeBuild project count to a Blueprint pipeline: zero-touch provisions a small, constant set — not one
       per asset or per stage.
 
     You can also validate the config locally before pushing:
