+++ docs/content/workshops/autopilot-pipeline/02-stages-approvals-naming.md
@@ -62,7 +62,7 @@ CloudFormation execution role — the wrapper threads them into synth and deploy
 
 ## Controlling the CloudFormation stack name
 
-v3 synthesizes the same `bin/` once per stage, so a bare `new MyStack(app, 'my-app')` deploys the same
+zero-touch synthesizes the same `bin/` once per stage, so a bare `new MyStack(app, 'my-app')` deploys the same
 name in every stage. Use the opt-in `stageStackName` helper to qualify it by stage:
 
 ```ts
