+++ docs/content/workshops/autopilot-pipeline/04-private-registry.md
