+++ docs/content/workshops/autopilot-pipeline/images/flat-pipeline.png
