+++ packages/@cdklabs/cdk-cicd-wrapper-cli/src/cmds/autopilot/DeployFromImage.ts
