+++ packages/@cdklabs/cdk-cicd-wrapper-cli/src/cmds/autopilot/SynthCommand.ts
