+++ packages/@cdklabs/cdk-cicd-wrapper-cli/src/index.ts
@@ -9,14 +9,14 @@ import complianceBucket from './cmds/ComplianceBucketCommand';
 import configure from './cmds/ConfigureCommand';
 import license from './cmds/LicenseCommand';
 import security from './cmds/SecurityCommand';
-import check from './cmds/zerotouch/CheckCommand';
-import deployCi from './cmds/zerotouch/DeployCiCommand';
-import deploy from './cmds/zerotouch/DeployCommand';
-import deploymentApp from './cmds/zerotouch/DeploymentAppCommand';
-import exec from './cmds/zerotouch/ExecCommand';
-import migrate from './cmds/zerotouch/MigrateCommand';
-import pipelineApp from './cmds/zerotouch/PipelineAppCommand';
-import synth from './cmds/zerotouch/SynthCommand';
+import check from './cmds/autopilot/CheckCommand';
+import deployCi from './cmds/autopilot/DeployCiCommand';
+import deploy from './cmds/autopilot/DeployCommand';
+import deploymentApp from './cmds/autopilot/DeploymentAppCommand';
+import exec from './cmds/autopilot/ExecCommand';
+import migrate from './cmds/autopilot/MigrateCommand';
+import pipelineApp from './cmds/autopilot/PipelineAppCommand';
+import synth from './cmds/autopilot/SynthCommand';
 import validate from './cmds/ValidateCommand';
 import { logger } from './utils/Logging';
