+++ packages/@cdklabs/cdk-cicd-wrapper-cli/test/autopilot/DriftCheck.test.ts
@@ -4,7 +4,7 @@
 import * as fs from 'fs';
 import * as os from 'os';
 import * as path from 'path';
-import { analyzeManifest, checkAssembly } from '../../src/cmds/zerotouch/DriftCheck';
+import { analyzeManifest, checkAssembly } from '../../src/cmds/autopilot/DriftCheck';
 
 function manifestWith(environment: string): any {
   return { artifacts: { TheStack: { type: 'aws:cloudformation:stack', environment } } };
